[0.4.1] String Lowercase 🌻
